refactor(local-model): extract LocalModelTransfer from repository
Split exclusive download/import job ownership, Range resume, SHA-256,
and progress/cancel into LocalModelTransfer so LocalModelRepository
keeps catalog, readiness, delete, and device performance only. No
intentional product behavior change; ILocalModelRepository unchanged.

Closes #480
